Moved LampManger to manager directory
